AutomationApp/src/it/unisa/diem/se/automationapp/event/MessageEvent.java
/*
 * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
 * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
 */
package it.unisa.diem.se.automationapp.event;

import com.fasterxml.jackson.annotation.JsonTypeName;

@JsonTypeName("message")
public class MessageEvent implements EventInterface{
    private String message;

    public MessageEvent() {
    }

    
    public MessageEvent(String message) {
        this.message = message;
    }

    @Override
    public String getMessage() {
        return this.message;
    }

    public void setMessage(String message) {
        this.message = message;
    }
    
}

